refactor(shared-session): coordinator 主体拆 ingress/canonical_blocks/identity 三子模块（2670→1280）
