+++ src-tauri/src/shared_runtime_coordinator.rs
@@ -8,13 +8,11 @@
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::sync::{Arc, Mutex};
 
-use serde_json::{json, Value};
+use serde_json::Value;
+#[cfg(test)]
+use serde_json::json;
 
 use crate::backend::events::AppServerEvent;
-use crate::engine::codex_prompt_service::{
-    extract_agent_message_snapshot_text, extract_codex_reasoning_delta, extract_codex_text_delta,
-    extract_turn_completed_text,
-};
 use crate::engine::events::EngineEvent;
 use crate::engine::EngineType;
 use crate::shared_event_log::canonical::assembler::{
@@ -24,6 +22,14 @@ use crate::shared_event_log::canonical::types::{
     ArtifactRef, CanonicalBlock, CanonicalOmission, OutcomeStatus, ProviderPrivateRef,
     ToolResultStatus, TurnExecutionSnapshot,
 };
+
+mod canonical_blocks;
+mod identity;
+mod ingress;
+
+pub(crate) use canonical_blocks::*;
+pub(crate) use identity::*;
+pub(crate) use ingress::*;
 
 const MAX_UNOWNED_EVENTS: usize = 512;
 const UNCLASSIFIED_RUNTIME_FAILURE_CODE: &str = "runtime_failure_unclassified";
@@ -155,7 +161,7 @@ pub(crate) struct SharedRuntimeReplayBatch {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct RuntimeIdentityKey {
+pub(crate) struct RuntimeIdentityKey {
     workspace_id: String,
     engine: EngineType,
     provider_runtime_key: String,
@@ -163,14 +169,14 @@ struct RuntimeIdentityKey {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct RuntimeScopeKey {
+pub(crate) struct RuntimeScopeKey {
     workspace_id: String,
     engine: EngineType,
     provider_runtime_key: String,
 }
 
 #[derive(Debug)]
-struct AttemptAccumulator {
+pub(crate) struct AttemptAccumulator {
     owner: SharedRuntimeAttemptOwner,
     assistant_blocks: Vec<CanonicalBlock>,
     tool_calls: Vec<RuntimeToolCall>,
@@ -187,7 +193,7 @@ struct AttemptAccumulator {
 }
 
 #[derive(Debug, Clone)]
-struct RuntimeIngress {
+pub(crate) struct RuntimeIngress {
     workspace_id: String,
     engine: EngineType,
     provider_runtime_key: String,
@@ -200,7 +206,7 @@ struct RuntimeIngress {
 }
 
 #[derive(Debug, Clone)]
-enum AccumulatorAction {
+pub(crate) enum AccumulatorAction {
     AssistantDelta(String),
     AssistantSnapshot(String),
     ReasoningDelta(String),
@@ -230,7 +236,7 @@ enum AccumulatorAction {
 }
 
 #[derive(Debug, Clone)]
-struct TerminalEvidence {
+pub(crate) struct TerminalEvidence {
     outcome: OutcomeStatus,
     error_code: Option<String>,
     error_message: Option<String>,
@@ -1265,1402 +1271,6 @@ impl CoordinatorState {
             // 让期间到达的 ingress 继续排队。下一轮空 drain 才原子清除。
             barrier_cleared: false,
         })
-    }
-}
-
-pub(crate) fn project_app_server_event_to_shared_owner(
-    event: &mut AppServerEvent,
-    owner: &SharedRuntimeAttemptOwner,
-) {
-    let requires_binding_recovery = owner.engine == EngineType::Claude
-        && is_missing_native_session_error(&event.message.to_string());
-    let native_thread_id = crate::backend::app_server::extract_thread_id(&event.message)
-        .filter(|thread_id| !thread_id.starts_with("shared:"))
-        .or_else(|| owner.native_session_id.clone());
-    // Read method before mutably borrowing params (same message object).
-    let method = event
-        .message
-        .get("method")
-        .and_then(|value| value.as_str())
-        .unwrap_or("")
-        .to_string();
-    // Control-plane methods are fail-closed on the frontend: params.turnId must
-    // equal sharedOwner.runtimeTurnId. Claude historically mapped
-    // requestUserInput.turnId to the assistant item id; force-align here so
-    // Shared AskUserQuestion / approval cards are not silently dropped.
-    let force_control_turn_identity = method == "item/tool/requestUserInput"
-        || method == "approval/request"
-        || method == "collaboration/modeBlocked"
-        || method.contains("requestApproval");
-    let params = event.message.as_object_mut().and_then(|message| {
-        message
-            .entry("params".to_string())
-            .or_insert_with(|| json!({}))
-            .as_object_mut()
-    });
-    let Some(params) = params else {
-        return;
-    };
-
-    params.insert(
-        "threadId".to_string(),
-        Value::String(owner.shared_thread_id.clone()),
-    );
-    params.insert(
-        "thread_id".to_string(),
-        Value::String(owner.shared_thread_id.clone()),
-    );
-    rewrite_nested_thread_identity(params.get_mut("turn"), &owner.shared_thread_id);
-    rewrite_nested_thread_identity(params.get_mut("thread"), &owner.shared_thread_id);
-
-    if let Some(native_thread_id) = native_thread_id.as_deref() {
-        params.insert(
-            "nativeThreadId".to_string(),
-            Value::String(native_thread_id.to_string()),
-        );
-        params.insert(
-            "native_thread_id".to_string(),
-            Value::String(native_thread_id.to_string()),
-        );
-    }
-    if let Some(runtime_turn_id) = owner.runtime_turn_id.as_deref() {
-        if force_control_turn_identity {
-            params.insert(
-                "turnId".to_string(),
-                Value::String(runtime_turn_id.to_string()),
-            );
-            params.insert(
-                "turn_id".to_string(),
-                Value::String(runtime_turn_id.to_string()),
-            );
-        } else {
-            params
-                .entry("turnId".to_string())
-                .or_insert_with(|| Value::String(runtime_turn_id.to_string()));
-            params
-                .entry("turn_id".to_string())
-                .or_insert_with(|| Value::String(runtime_turn_id.to_string()));
-        }
-    }
-    params.insert(
-        "sharedOwner".to_string(),
-        json!({
-            "sharedSessionId": owner.shared_session_id,
-            "sharedThreadId": owner.shared_thread_id,
-            "providerRuntimeKey": owner.provider_runtime_key,
-            "nativeThreadId": native_thread_id,
-            "runtimeTurnId": owner.runtime_turn_id,
-            "logicalTurnId": owner.logical_turn_id,
-            "attemptId": owner.attempt_id,
-            "bindingKey": owner.binding_key,
-            "bindingOperationId": owner.binding_operation_id,
-            "engine": engine_token(owner.engine),
-            "executionTargetSnapshot": execution_target_snapshot_wire(
-                &owner.execution_target_snapshot
-            ),
-        }),
-    );
-    if requires_binding_recovery {
-        params.insert(
-            "sharedRecoveryReason".to_string(),
-            Value::String("native-session-not-found".to_string()),
-        );
-    }
-}
-
-fn execution_target_snapshot_wire(snapshot: &TurnExecutionSnapshot) -> Value {
-    json!({
-        "engine": snapshot.engine,
-        "providerProfileId": snapshot.provider_profile_id,
-        "modelCatalogEntryId": snapshot.model_catalog_entry_id,
-        "model": snapshot.model,
-        "reasoning": snapshot.reasoning,
-        "providerProfileNameSnapshot": snapshot.provider_profile_name_snapshot,
-        "providerProfileSource": snapshot.provider_profile_source,
-        "runtimeCapabilityFingerprint": snapshot.runtime_capability_fingerprint,
-    })
-}
-
-fn rewrite_nested_thread_identity(value: Option<&mut Value>, shared_thread_id: &str) {
-    let Some(object) = value.and_then(Value::as_object_mut) else {
-        return;
-    };
-    object.insert(
-        "threadId".to_string(),
-        Value::String(shared_thread_id.to_string()),
-    );
-    object.insert(
-        "thread_id".to_string(),
-        Value::String(shared_thread_id.to_string()),
-    );
-}
-
-fn normalize_engine_ingress(
-    provider_runtime_key: &str,
-    engine: EngineType,
-    runtime_turn_id: Option<&str>,
-    native_session_id: Option<&str>,
-    event: &EngineEvent,
-) -> RuntimeIngress {
-    let mut actions = Vec::new();
-    let mut suppress_agent_event = false;
-    match event {
-        EngineEvent::SessionStarted { session_id, .. } => {
-            return RuntimeIngress {
-                workspace_id: event.workspace_id().to_string(),
-                engine,
-                provider_runtime_key: provider_runtime_key.to_string(),
-                runtime_turn_id: normalize_identity(runtime_turn_id).map(str::to_string),
-                native_session_id: normalize_native_session_identity(
-                    engine,
-                    Some(provider_runtime_key),
-                    Some(session_id.as_str()),
-                )
-                .or_else(|| {
-                    normalize_native_session_identity(
-                        engine,
-                        Some(provider_runtime_key),
-                        native_session_id,
-                    )
-                }),
-                is_session_started: true,
-                actions,
-                agent_event: Some(event.clone()),
-                replay_app_server_events: Vec::new(),
-            };
-        }
-        EngineEvent::TurnStarted { turn_id, .. } => {
-            return RuntimeIngress {
-                workspace_id: event.workspace_id().to_string(),
-                engine,
-                provider_runtime_key: provider_runtime_key.to_string(),
-                runtime_turn_id: normalize_identity(Some(turn_id.as_str()))
-                    .or_else(|| normalize_identity(runtime_turn_id))
-                    .map(str::to_string),
-                native_session_id: normalize_native_session_identity(
-                    engine,
-                    Some(provider_runtime_key),
-                    native_session_id,
-                ),
-                is_session_started: false,
-                actions,
-                agent_event: Some(event.clone()),
-                replay_app_server_events: Vec::new(),
-            };
-        }
-        EngineEvent::TextDelta { text, .. } => {
-            actions.push(AccumulatorAction::AssistantDelta(text.clone()));
-        }
-        EngineEvent::ReasoningDelta { text, .. } => {
-            actions.push(AccumulatorAction::ReasoningDelta(text.clone()));
-        }
-        EngineEvent::ToolStarted {
-            tool_id,
-            tool_name,
-            input,
-            ..
-        } => actions.push(AccumulatorAction::ToolStarted {
-            tool_id: tool_id.clone(),
-            tool_name: tool_name.clone(),
-            input: input.clone(),
-        }),
-        EngineEvent::ToolInputUpdated {
-            tool_id,
-            tool_name,
-            input,
-            ..
-        } => actions.push(AccumulatorAction::ToolInputUpdated {
-            tool_id: tool_id.clone(),
-            tool_name: tool_name.clone(),
-            input: input.clone(),
-        }),
-        EngineEvent::ToolOutputDelta { tool_id, delta, .. } => {
-            actions.push(AccumulatorAction::ToolOutputDelta {
-                tool_id: tool_id.clone(),
-                delta: delta.clone(),
-            });
-        }
-        EngineEvent::ToolCompleted {
-            tool_id,
-            tool_name,
-            output,
-            error,
-            ..
-        } => actions.push(AccumulatorAction::ToolCompleted {
-            tool_id: tool_id.clone(),
-            tool_name: tool_name.clone(),
-            output: output.clone(),
-            error: error.clone(),
-        }),
-        EngineEvent::TurnCompleted { result, .. } => {
-            let result = result.as_ref();
-            actions.push(AccumulatorAction::Terminal(TerminalEvidence {
-                outcome: completion_outcome(result),
-                error_code: value_string_by_aliases(result, &["errorCode", "error_code", "code"]),
-                error_message: value_string_by_aliases(
-                    result,
-                    &["errorMessage", "error_message", "error"],
-                ),
-                stop_reason: value_string_by_aliases(
-                    result,
-                    &["stopReason", "stop_reason", "reason"],
-                ),
-                fallback_text: crate::engine::commands::extract_turn_result_text(result),
-                artifacts: result
-                    .map(extract_explicit_artifact_refs)
-                    .unwrap_or_default(),
-                provider_private_refs: deserialize_vec_by_aliases(
-                    result,
-                    &["providerPrivateRefs", "provider_private_refs"],
-                ),
-                omissions: deserialize_vec_by_aliases(result, &["omissions"]),
-            }));
-        }
-        EngineEvent::TurnError { error, code, .. } => {
-            actions.push(AccumulatorAction::Terminal(TerminalEvidence {
-                outcome: OutcomeStatus::Failed,
-                error_code: code.clone(),
-                error_message: Some(error.clone()),
-                stop_reason: None,
-                fallback_text: None,
-                artifacts: Vec::new(),
-                provider_private_refs: Vec::new(),
-                omissions: Vec::new(),
-            }));
-        }
-        EngineEvent::Raw { data, .. } => {
-            if engine == EngineType::Claude {
-                if let Some(echo) = extract_claude_replay_echo(data) {
-                    actions.push(AccumulatorAction::ContextEcho(echo));
-                    // replay echo 是 transport-level ACK，不是用户输入，也不是 assistant
-                    // content。禁止进入 AgentEventBus/history/UI raw fan-out。
-                    suppress_agent_event = true;
-                }
-                if let Some(terminal) = claude_result_terminal_evidence(data) {
-                    // Claude CLI 的 `result` packet 是业务回合已经结束的 typed
-                    // evidence；后续 stdout/stderr drain 与 process reap 只是 Runtime
-                    // cleanup。Native Claude 仍等待 canonical TurnCompleted，Shared
-                    // attempt 则必须在这里先收口，不能让清理延迟继续占用 Stop/UI lock。
-                    actions.push(AccumulatorAction::Terminal(terminal));
-                }
-            }
-            let artifacts = extract_explicit_artifact_refs(data);
-            if !artifacts.is_empty() {
-                actions.push(AccumulatorAction::Artifacts(artifacts));
-            }
-        }
-        _ => {}
-    }
-    RuntimeIngress {
-        workspace_id: event.workspace_id().to_string(),
-        engine,
-        provider_runtime_key: provider_runtime_key.to_string(),
-        runtime_turn_id: normalize_identity(runtime_turn_id).map(str::to_string),
-        native_session_id: normalize_native_session_identity(
-            engine,
-            Some(provider_runtime_key),
-            native_session_id,
-        ),
-        is_session_started: false,
-        actions,
-        agent_event: (!suppress_agent_event).then(|| event.clone()),
-        replay_app_server_events: Vec::new(),
-    }
-}
-
-fn claude_result_terminal_evidence(data: &Value) -> Option<TerminalEvidence> {
-    let event_type = data
-        .get("type")
-        .and_then(Value::as_str)
-        .unwrap_or_default()
-        .trim();
-    if !event_type.eq_ignore_ascii_case("result") {
-        return None;
-    }
-
-    let subtype = data
-        .get("subtype")
-        .and_then(Value::as_str)
-        .unwrap_or_default()
-        .trim()
-        .to_ascii_lowercase();
-    let terminal_reason = value_string_by_aliases(
-        Some(data),
-        &[
-            "terminalReason",
-            "terminal_reason",
-            "stopReason",
-            "stop_reason",
-        ],
-    );
-    let is_failed = data
-        .get("is_error")
-        .or_else(|| data.get("isError"))
-        .and_then(Value::as_bool)
-        .unwrap_or(false)
-        || matches!(
-            subtype.as_str(),
-            "error" | "failed" | "failure" | "error_during_execution"
-        )
-        || subtype.starts_with("error_");
-
-    let fallback_text = (!is_failed)
-        .then(|| crate::engine::commands::extract_turn_result_text(Some(data)))
-        .flatten();
-    let result_text = crate::engine::commands::extract_turn_result_text(Some(data));
-    let error_message = if is_failed {
-        value_string_by_aliases(Some(data), &["errorMessage", "error_message", "message"])
-            .or_else(|| {
-                data.get("error")
-                    .and_then(|error| value_string_by_aliases(Some(error), &["message"]))
-            })
-            .or(result_text)
-    } else {
-        None
-    };
-
-    Some(TerminalEvidence {
-        outcome: if is_failed {
-            OutcomeStatus::Failed
-        } else {
-            OutcomeStatus::Completed
-        },
-        error_code: value_string_by_aliases(
-            Some(data),
-            &[
-                "apiErrorStatus",
-                "api_error_status",
-                "errorCode",
-                "error_code",
-                "code",
-            ],
-        ),
-        error_message,
-        stop_reason: terminal_reason,
-        fallback_text,
-        artifacts: extract_explicit_artifact_refs(data),
-        provider_private_refs: deserialize_vec_by_aliases(
-            Some(data),
-            &["providerPrivateRefs", "provider_private_refs"],
-        ),
-        omissions: deserialize_vec_by_aliases(Some(data), &["omissions"]),
-    })
-}
-
-/// Claude `--replay-user-messages` 回显的 Shared context marker 仅用于强 ACK。
-/// 即使 runtime identity 尚未完成绑定，也必须在普通 UI/history fan-out 前过滤。
-pub(crate) fn is_internal_shared_context_replay_event(event: &EngineEvent) -> bool {
-    matches!(
-        event,
-        EngineEvent::Raw {
-            engine: EngineType::Claude,
-            data,
-            ..
-        } if extract_claude_replay_echo(data).is_some()
-    )
-}
-
-fn normalize_codex_ingress(
-    provider_runtime_key: &str,
-    workspace_id: &str,
-    event: &Value,
-) -> RuntimeIngress {
-    let method = event
-        .get("method")
-        .and_then(Value::as_str)
-        .unwrap_or_default();
-    let params = event.get("params").unwrap_or(&Value::Null);
-    let runtime_turn_id = crate::backend::app_server::extract_turn_id(event);
-    let native_session_id = crate::backend::app_server::extract_thread_id(event);
-    let mut actions = Vec::new();
-
-    let agent_event = if let Some(text) = extract_codex_text_delta(event) {
-        actions.push(AccumulatorAction::AssistantDelta(text.clone()));
-        Some(EngineEvent::TextDelta {
-            workspace_id: workspace_id.to_string(),
-            text,
-        })
-    } else if let Some(text) = extract_codex_reasoning_delta(event) {
-        actions.push(AccumulatorAction::ReasoningDelta(text.clone()));
-        Some(EngineEvent::ReasoningDelta {
-            workspace_id: workspace_id.to_string(),
-            text,
-        })
-    } else if let Some(text) = extract_agent_message_snapshot_text(event) {
-        actions.push(AccumulatorAction::AssistantSnapshot(text));
-        Some(EngineEvent::Raw {
-            workspace_id: workspace_id.to_string(),
-            engine: EngineType::Codex,
-            data: event.clone(),
-        })
-    } else {
-        normalize_codex_structured_event(workspace_id, method, params, event, &mut actions)
-    };
-
-    let artifacts = extract_explicit_artifact_refs(params);
-    if !artifacts.is_empty() {
-        actions.push(AccumulatorAction::Artifacts(artifacts));
-    }
-
-    RuntimeIngress {
-        workspace_id: workspace_id.to_string(),
-        engine: EngineType::Codex,
-        provider_runtime_key: provider_runtime_key.to_string(),
-        runtime_turn_id,
-        native_session_id,
-        is_session_started: method == "thread/started",
-        actions,
-        agent_event,
-        replay_app_server_events: vec![AppServerEvent {
-            workspace_id: workspace_id.to_string(),
-            message: event.clone(),
-        }],
-    }
-}
-
-fn normalize_codex_structured_event(
-    workspace_id: &str,
-    method: &str,
-    params: &Value,
-    event: &Value,
-    actions: &mut Vec<AccumulatorAction>,
-) -> Option<EngineEvent> {
-    match method {
-        "turn/started" => {
-            let turn_id = crate::backend::app_server::extract_turn_id(event).unwrap_or_default();
-            Some(EngineEvent::TurnStarted {
-                workspace_id: workspace_id.to_string(),
-                turn_id,
-            })
-        }
-        "thread/started" => {
-            let session_id =
-                crate::backend::app_server::extract_thread_id(event).unwrap_or_default();
-            Some(EngineEvent::SessionStarted {
-                workspace_id: workspace_id.to_string(),
-                session_id,
-                engine: EngineType::Codex,
-                turn_id: crate::backend::app_server::extract_turn_id(event),
-            })
-        }
-        "item/started" | "item/updated" | "item/completed" => {
-            normalize_codex_item_event(workspace_id, method, params, actions)
-        }
-        "item/toolStart" => {
-            let tool_id = value_string_by_aliases(Some(params), &["toolId", "tool_id", "id"])
-                .unwrap_or_else(|| "unknown-tool".to_string());
-            let tool_name =
-                value_string_by_aliases(Some(params), &["toolName", "tool_name", "name"])
-                    .unwrap_or_else(|| tool_id.clone());
-            let input = value_by_aliases(params, &["input", "arguments"]).cloned();
-            actions.push(AccumulatorAction::ToolStarted {
-                tool_id: tool_id.clone(),
-                tool_name: tool_name.clone(),
-                input: input.clone(),
-            });
-            Some(EngineEvent::ToolStarted {
-                workspace_id: workspace_id.to_string(),
-                tool_id,
-                tool_name,
-                input,
-            })
-        }
-        "item/toolComplete" => {
-            let tool_id = value_string_by_aliases(Some(params), &["toolId", "tool_id", "id"])
-                .unwrap_or_else(|| "unknown-tool".to_string());
-            let tool_name =
-                value_string_by_aliases(Some(params), &["toolName", "tool_name", "name"]);
-            let output = value_by_aliases(params, &["output", "result"]).cloned();
-            let error = value_string_by_aliases(Some(params), &["error", "errorMessage"]);
-            actions.push(AccumulatorAction::ToolCompleted {
-                tool_id: tool_id.clone(),
-                tool_name: tool_name.clone(),
-                output: output.clone(),
-                error: error.clone(),
-            });
-            Some(EngineEvent::ToolCompleted {
-                workspace_id: workspace_id.to_string(),
-                tool_id,
-                tool_name,
-                output,
-                error,
-            })
-        }
-        "turn/completed" => {
-            let fallback_text = extract_turn_completed_text(event);
-            let evidence = terminal_evidence_from_value(params, fallback_text, false);
-            actions.push(AccumulatorAction::Terminal(evidence));
-            Some(EngineEvent::TurnCompleted {
-                workspace_id: workspace_id.to_string(),
-                result: Some(params.clone()),
-            })
-        }
-        "turn/error" | "runtime/ended" => {
-            let evidence = terminal_evidence_from_value(params, None, true);
-            let error = evidence
-                .error_message
-                .clone()
-                .unwrap_or_else(|| "Codex runtime turn failed".to_string());
-            let code = evidence.error_code.clone();
-            actions.push(AccumulatorAction::Terminal(evidence));
-            Some(EngineEvent::TurnError {
-                workspace_id: workspace_id.to_string(),
-                error,
-                code,
-            })
-        }
-        "error"
-            if !params
-                .get("willRetry")
-                .or_else(|| params.get("will_retry"))
-                .and_then(Value::as_bool)
-                .unwrap_or(false) =>
-        {
-            let evidence = terminal_evidence_from_value(params, None, true);
-            let error = evidence
-                .error_message
-                .clone()
-                .unwrap_or_else(|| "Codex runtime request failed".to_string());
-            let code = evidence.error_code.clone();
-            actions.push(AccumulatorAction::Terminal(evidence));
-            Some(EngineEvent::TurnError {
-                workspace_id: workspace_id.to_string(),
-                error,
-                code,
-            })
-        }
-        _ => Some(EngineEvent::Raw {
-            workspace_id: workspace_id.to_string(),
-            engine: EngineType::Codex,
-            data: event.clone(),
-        }),
-    }
-}
-
-fn normalize_codex_item_event(
-    workspace_id: &str,
-    method: &str,
-    params: &Value,
-    actions: &mut Vec<AccumulatorAction>,
-) -> Option<EngineEvent> {
-    let item = params.get("item")?;
-    if is_assistant_or_reasoning_item(item) {
-        return Some(EngineEvent::Raw {
-            workspace_id: workspace_id.to_string(),
-            engine: EngineType::Codex,
-            data: json!({ "method": method, "params": params }),
-        });
-    }
-    let item_type = value_string_by_aliases(Some(item), &["type", "kind"]).unwrap_or_default();
-    if !is_tool_item_type(&item_type) {
-        return Some(EngineEvent::Raw {
-            workspace_id: workspace_id.to_string(),
-            engine: EngineType::Codex,
-            data: json!({ "method": method, "params": params }),
-        });
-    }
-    let tool_id = value_string_by_aliases(Some(item), &["id", "toolId", "tool_id"])
-        .unwrap_or_else(|| "unknown-tool".to_string());
-    // Prefer explicit tool name; custom_tool_call uses `name` (e.g. apply_patch).
-    // Fall back to item type (e.g. "fileChange") for canvas classifiers.
-    let tool_name = value_string_by_aliases(
-        Some(item),
-        &["tool", "toolName", "tool_name", "name", "title"],
-    )
-    .unwrap_or_else(|| item_type.clone());
-    // Codex fileChange puts paths/diffs on `changes[]`, not `arguments`/`input`.
-    // Pack both so SharedProjector can rebuild ConversationItem.changes.
-    let input = extract_codex_tool_payload(item);
-    if method == "item/started" {
-        actions.push(AccumulatorAction::ToolStarted {
-            tool_id: tool_id.clone(),
-            tool_name: tool_name.clone(),
-            input: input.clone(),
-        });
-        return Some(EngineEvent::ToolStarted {
-            workspace_id: workspace_id.to_string(),
-            tool_id,
-            tool_name,
-            input,
-        });
-    }
-    if method == "item/updated" {
-        actions.push(AccumulatorAction::ToolInputUpdated {
-            tool_id: tool_id.clone(),
-            tool_name: Some(tool_name.clone()),
-            input: input.clone(),
-        });
-        return Some(EngineEvent::ToolInputUpdated {
-            workspace_id: workspace_id.to_string(),
-            tool_id,
-            tool_name: Some(tool_name),
-            input,
-        });
-    }
-
-    // Completed snapshots often carry the final `changes[]` only at this step.
-    if let Some(payload) = input.clone() {
-        actions.push(AccumulatorAction::ToolInputUpdated {
-            tool_id: tool_id.clone(),
-            tool_name: Some(tool_name.clone()),
-            input: Some(payload),
-        });
-    }
-    let output = value_by_aliases(item, &["result", "output", "aggregatedOutput"])
-        .or_else(|| value_by_aliases(params, &["result", "output"]))
-        .cloned();
-    let error = value_string_by_aliases(Some(item), &["error", "errorMessage"])
-        .or_else(|| value_string_by_aliases(Some(params), &["error", "errorMessage"]));
-    actions.push(AccumulatorAction::ToolCompleted {
-        tool_id: tool_id.clone(),
-        tool_name: Some(tool_name.clone()),
-        output: output.clone(),
-        error: error.clone(),
-    });
-    Some(EngineEvent::ToolCompleted {
-        workspace_id: workspace_id.to_string(),
-        tool_id,
-        tool_name: Some(tool_name),
-        output,
-        error,
-    })
-}
-
-/// Build a portable tool payload for Shared canonical storage.
-///
-/// Codex `fileChange` items put path/diff on `changes[]` (not `arguments`).
-/// Codex `apply_patch` often arrives as `custom_tool_call` with a raw patch string
-/// in `input`. Both must be packed or history cannot rebuild the file-edit scene.
-fn extract_codex_tool_payload(item: &Value) -> Option<Value> {
-    let mut object = serde_json::Map::new();
-
-    match value_by_aliases(item, &["arguments", "input"]) {
-        Some(Value::Object(map)) => {
-            for (key, value) in map {
-                object.insert(key.clone(), value.clone());
-            }
-        }
-        Some(Value::String(text)) => {
-            let trimmed = text.trim();
-            if !trimmed.is_empty() {
-                object.insert("input".to_string(), Value::String(trimmed.to_string()));
-                // Preserve patch-shaped strings under an explicit key for projection.
-                if trimmed.contains("*** Begin Patch") || trimmed.contains("*** Update File:") {
-                    object.insert("patch".to_string(), Value::String(trimmed.to_string()));
-                }
-            }
-        }
-        Some(value) if !value.is_null() => {
-            object.insert("input".to_string(), value.clone());
-        }
-        _ => {}
-    }
-
-    if let Some(changes) = item.get("changes") {
-        if changes.as_array().is_some_and(|rows| !rows.is_empty()) {
-            object.insert("changes".to_string(), changes.clone());
-        }
-    }
-
-    // custom_tool_call / function_call name (apply_patch, shell, …)
-    if let Some(name) =
-        value_string_by_aliases(Some(item), &["name", "tool", "toolName", "tool_name"])
-    {
-        let trimmed = name.trim();
-        if !trimmed.is_empty() {
-            object.insert("name".to_string(), Value::String(trimmed.to_string()));
-        }
-    }
-
-    if let Some(title) = item.get("title").and_then(Value::as_str) {
-        let trimmed = title.trim();
-        if !trimmed.is_empty() {
-            object.insert("title".to_string(), Value::String(trimmed.to_string()));
-        }
-    }
-
-    // commandExecution-shaped fields. Codex often sends `command` as a string[] argv
-    // (e.g. ["cat","README.md"] or apply_patch + patch body). We must join argv into a
-    // single string or Shared history loses the command text and cannot promote
-    // apply_patch → fileChange.
-    for key in ["cwd", "description"] {
-        if let Some(Value::String(text)) = item.get(key) {
-            let trimmed = text.trim();
-            if !trimmed.is_empty() {
-                object.insert(key.to_string(), Value::String(trimmed.to_string()));
-            }
-        }
-    }
-    if let Some(command) = coerce_command_field(item.get("command").or_else(|| item.get("cmd"))) {
-        let looks_like_patch = command.contains("*** Begin Patch")
-            || command.contains("*** Update File:")
-            || command.to_ascii_lowercase().contains("apply_patch");
-        object.insert("command".to_string(), Value::String(command.clone()));
-        if looks_like_patch {
-            object.insert("patch".to_string(), Value::String(command));
-        }
-    }
-
-    if object.is_empty() {
-        None
-    } else {
-        Some(Value::Object(object))
-    }
-}
-
-/// Normalize Codex command field: string as-is, string[] joined with spaces.
-fn coerce_command_field(value: Option<&Value>) -> Option<String> {
-    match value {
-        Some(Value::String(text)) => {
-            let trimmed = text.trim();
-            if trimmed.is_empty() {
-                None
-            } else {
-                Some(trimmed.to_string())
-            }
-        }
-        Some(Value::Array(parts)) => {
-            let joined = parts
-                .iter()
-                .filter_map(|part| part.as_str().map(str::trim))
-                .filter(|part| !part.is_empty())
-                .collect::<Vec<_>>()
-                .join(" ");
-            if joined.is_empty() {
-                None
-            } else {
-                Some(joined)
-            }
-        }
-        _ => None,
-    }
-}
-
-/// Merge tool argument JSON summaries. Object keys from `incoming` win; non-JSON
-/// strings fall back to last-write-wins (preserves prior string when incoming empty).
-fn merge_tool_arguments_summary(existing: Option<&str>, incoming: &str) -> String {
-    let incoming = incoming.trim();
-    if incoming.is_empty() {
-        return existing.unwrap_or("").to_string();
-    }
-    let Some(existing) = existing.map(str::trim).filter(|text| !text.is_empty()) else {
-        return incoming.to_string();
-    };
-    let Ok(Value::Object(mut base)) = serde_json::from_str::<Value>(existing) else {
-        return incoming.to_string();
-    };
-    let Ok(Value::Object(patch)) = serde_json::from_str::<Value>(incoming) else {
-        return incoming.to_string();
-    };
-    for (key, value) in patch {
-        base.insert(key, value);
-    }
-    serde_json::to_string(&Value::Object(base)).unwrap_or_else(|_| incoming.to_string())
-}
-
-fn terminal_evidence_from_value(
-    value: &Value,
-    fallback_text: Option<String>,
-    force_failed: bool,
-) -> TerminalEvidence {
-    let outcome = if force_failed {
-        OutcomeStatus::Failed
-    } else {
-        completion_outcome(Some(value))
-    };
-    TerminalEvidence {
-        outcome,
-        error_code: value_string_by_aliases(
-            Some(value),
-            &[
-                "errorCode",
-                "error_code",
-                "code",
-                "reasonCode",
-                "reason_code",
-            ],
-        )
-        .or_else(|| {
-            value.get("error").and_then(|error| {
-                value_string_by_aliases(
-                    Some(error),
-                    &[
-                        "errorCode",
-                        "error_code",
-                        "code",
-                        "reasonCode",
-                        "reason_code",
-                    ],
-                )
-            })
-        }),
-        error_message: value_string_by_aliases(
-            Some(value),
-            &["errorMessage", "error_message", "message"],
-        )
-        .or_else(|| {
-            value
-                .get("error")
-                .and_then(|error| value_string_by_aliases(Some(error), &["message"]))
-        }),
-        stop_reason: value_string_by_aliases(Some(value), &["stopReason", "stop_reason", "reason"]),
-        fallback_text,
-        artifacts: extract_explicit_artifact_refs(value),
-        provider_private_refs: deserialize_vec_by_aliases(
-            Some(value),
-            &["providerPrivateRefs", "provider_private_refs"],
-        ),
-        omissions: deserialize_vec_by_aliases(Some(value), &["omissions"]),
-    }
-}
-
-fn push_assistant_block(blocks: &mut Vec<CanonicalBlock>, next: CanonicalBlock) {
-    match (blocks.last_mut(), next) {
-        (Some(CanonicalBlock::Text { text }), CanonicalBlock::Text { text: delta }) => {
-            text.push_str(&delta);
-        }
-        (Some(CanonicalBlock::Reasoning { text }), CanonicalBlock::Reasoning { text: delta }) => {
-            text.push_str(&delta);
-        }
-        (_, next) => blocks.push(next),
-    }
-}
-
-fn canonical_block_text(block: &CanonicalBlock) -> (&'static str, &str) {
-    match block {
-        CanonicalBlock::Text { text } => ("text", text),
-        CanonicalBlock::Reasoning { text } => ("reasoning", text),
-        _ => ("other", ""),
-    }
-}
-
-fn matching_block_text(blocks: &[CanonicalBlock], kind: &str) -> String {
-    blocks
-        .iter()
-        .filter_map(|block| {
-            let (block_kind, text) = canonical_block_text(block);
-            (block_kind == kind).then_some(text)
-        })
-        .collect()
-}
-
-fn is_full_claude_observation(text: &str) -> bool {
-    text.chars()
-        .filter(|character| !character.is_whitespace())
-        .count()
-        >= CLAUDE_FULL_OBSERVATION_MIN_CHARS
-}
-
-/// Claude provider adapters may expose a growing full snapshot on more than one protocol
-/// surface. This merge is intentionally scoped to the Shared accumulator: Native Claude keeps
-/// its existing reducer normalization and Codex keeps delta append semantics.
-fn merge_claude_full_observation(blocks: &mut Vec<CanonicalBlock>, next: CanonicalBlock) {
-    let (kind, incoming) = canonical_block_text(&next);
-    if kind == "other" || incoming.trim().is_empty() {
-        push_assistant_block(blocks, next);
-        return;
-    }
-
-    let existing = matching_block_text(blocks, kind);
-    if existing.is_empty() || !is_full_claude_observation(&existing) {
-        push_assistant_block(blocks, next);
-        return;
-    }
-    if incoming == existing || existing.starts_with(incoming) {
-        return;
-    }
-    if let Some(suffix) = incoming.strip_prefix(&existing) {
-        let replay_trimmed = suffix.trim_start();
-        if let Some(after_replay) = replay_trimmed.strip_prefix(&existing) {
-            if !after_replay.is_empty() {
-                let replay_free = match kind {
-                    "text" => CanonicalBlock::Text {
-                        text: after_replay.to_string(),
-                    },
-                    "reasoning" => CanonicalBlock::Reasoning {
-                        text: after_replay.to_string(),
-                    },
-                    _ => unreachable!("canonical block kind checked above"),
-                };
-                push_assistant_block(blocks, replay_free);
-            }
-            return;
-        }
-        if !suffix.is_empty() {
-            let incremental_suffix = match kind {
-                "text" => CanonicalBlock::Text {
-                    text: suffix.to_string(),
-                },
-                "reasoning" => CanonicalBlock::Reasoning {
-                    text: suffix.to_string(),
-                },
-                _ => unreachable!("canonical block kind checked above"),
-            };
-            push_assistant_block(blocks, incremental_suffix);
-        }
-        return;
-    }
-
-    push_assistant_block(blocks, next);
-}
-
-fn merge_claude_complete_assistant_text(blocks: &mut Vec<CanonicalBlock>, complete_text: String) {
-    let existing_text = matching_block_text(blocks, "text");
-    if is_full_claude_observation(&existing_text) {
-        if complete_text == existing_text || existing_text.starts_with(&complete_text) {
-            return;
-        }
-        if let Some(suffix) = complete_text.strip_prefix(&existing_text) {
-            if suffix.trim_start().starts_with(&existing_text) {
-                return;
-            }
-        }
-    }
-    merge_complete_assistant_text(blocks, complete_text);
-}
-
-/// Snapshot/terminal text 是累计完成证据。只做单调补全；无前缀关系时保留独立
-/// Text block，禁止用猜测覆盖已观察到的 streamed content。
-fn merge_complete_assistant_text(blocks: &mut Vec<CanonicalBlock>, complete_text: String) {
-    if complete_text.trim().is_empty() {
-        return;
-    }
-
-    let existing_text = blocks
-        .iter()
-        .filter_map(|block| match block {
-            CanonicalBlock::Text { text } => Some(text.as_str()),
-            _ => None,
-        })
-        .collect::<String>();
-    if existing_text.is_empty() {
-        blocks.push(CanonicalBlock::Text {
-            text: complete_text,
-        });
-        return;
-    }
-    if complete_text == existing_text || existing_text.starts_with(&complete_text) {
-        return;
-    }
-    if let Some(suffix) = complete_text.strip_prefix(&existing_text) {
-        if let Some(CanonicalBlock::Text { text }) = blocks
-            .iter_mut()
-            .rev()
-            .find(|block| matches!(block, CanonicalBlock::Text { .. }))
-        {
-            text.push_str(suffix);
-        }
-        return;
-    }
-    if blocks
-        .iter()
-        .any(|block| matches!(block, CanonicalBlock::Text { text } if text == &complete_text))
-    {
-        return;
-    }
-
-    blocks.push(CanonicalBlock::Text {
-        text: complete_text,
-    });
-}
-
-fn upsert_tool_result(results: &mut Vec<RuntimeToolResult>, result: RuntimeToolResult) {
-    if let Some(existing) = results
-        .iter_mut()
-        .find(|existing| existing.tool_call_id == result.tool_call_id)
-    {
-        *existing = result;
-    } else {
-        results.push(result);
-    }
-}
-
-fn stringify_json_value(value: &Value) -> String {
-    value
-        .as_str()
-        .map(str::to_string)
-        .unwrap_or_else(|| serde_json::to_string(value).unwrap_or_default())
-}
-
-fn extract_explicit_artifact_refs(value: &Value) -> Vec<ArtifactRef> {
-    fn walk(value: &Value, refs: &mut Vec<ArtifactRef>) {
-        match value {
-            Value::Array(items) => items.iter().for_each(|item| walk(item, refs)),
-            Value::Object(object) => {
-                for (key, child) in object {
-                    if matches!(
-                        key.as_str(),
-                        "artifactRef" | "artifact_ref" | "artifactRefs" | "artifact_refs"
-                    ) {
-                        match child {
-                            Value::Array(items) => {
-                                for item in items {
-                                    if let Ok(artifact) =
-                                        serde_json::from_value::<ArtifactRef>(item.clone())
-                                    {
-                                        refs.push(artifact);
-                                    }
-                                }
-                            }
-                            _ => {
-                                if let Ok(artifact) =
-                                    serde_json::from_value::<ArtifactRef>(child.clone())
-                                {
-                                    refs.push(artifact);
-                                }
-                            }
-                        }
-                    } else {
-                        walk(child, refs);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    let mut refs = Vec::new();
-    walk(value, &mut refs);
-    let mut unique = Vec::new();
-    extend_unique_artifacts(&mut unique, refs);
-    unique
-}
-
-fn extract_claude_replay_echo(data: &Value) -> Option<String> {
-    let is_replay = data
-        .get("isReplay")
-        .or_else(|| data.get("is_replay"))
-        .and_then(Value::as_bool)
-        == Some(true);
-    if !is_replay {
-        return None;
-    }
-    fn collect_text(value: &Value, output: &mut String) {
-        match value {
-            Value::String(text) => output.push_str(text),
-            Value::Array(items) => {
-                for item in items {
-                    collect_text(item, output);
-                }
-            }
-            Value::Object(object) => {
-                for key in ["content", "text", "message"] {
-                    if let Some(value) = object.get(key) {
-                        collect_text(value, output);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    let mut echo = String::new();
-    collect_text(data.get("message").unwrap_or(data), &mut echo);
-    echo.contains("MOSSX_CONTEXT_PACKAGE:")
-        .then_some(echo)
-        .filter(|value| !value.is_empty())
-}
-
-fn extend_unique_artifacts(target: &mut Vec<ArtifactRef>, values: Vec<ArtifactRef>) {
-    for value in values {
-        if !target.iter().any(|existing| {
-            existing.artifact_id == value.artifact_id && existing.sha256 == value.sha256
-        }) {
-            target.push(value);
-        }
-    }
-}
-
-fn extend_unique_private_refs(
-    target: &mut Vec<ProviderPrivateRef>,
-    values: Vec<ProviderPrivateRef>,
-) {
-    for value in values {
-        if !target
-            .iter()
-            .any(|existing| existing.ref_id == value.ref_id)
-        {
-            target.push(value);
-        }
-    }
-}
-
-fn extend_unique_omissions(target: &mut Vec<CanonicalOmission>, values: Vec<CanonicalOmission>) {
-    for value in values {
-        if !target.iter().any(|existing| {
-            existing.category == value.category
-                && existing.reason == value.reason
-                && existing.retrievable_ref == value.retrievable_ref
-        }) {
-            target.push(value);
-        }
-    }
-}
-
-fn deserialize_vec_by_aliases<T: serde::de::DeserializeOwned>(
-    value: Option<&Value>,
-    aliases: &[&str],
-) -> Vec<T> {
-    let Some(value) = value else {
-        return Vec::new();
-    };
-    for alias in aliases {
-        if let Some(candidate) = value.get(*alias) {
-            if let Ok(values) = serde_json::from_value::<Vec<T>>(candidate.clone()) {
-                return values;
-            }
-        }
-    }
-    Vec::new()
-}
-
-fn completion_outcome(value: Option<&Value>) -> OutcomeStatus {
-    let status =
-        value_string_by_aliases(value, &["status", "outcome", "stopReason", "stop_reason"])
-            .or_else(|| {
-                value.and_then(|root| {
-                    ["turn", "result"].iter().find_map(|key| {
-                        root.get(*key).and_then(|nested| {
-                            value_string_by_aliases(
-                                Some(nested),
-                                &["status", "outcome", "stopReason", "stop_reason"],
-                            )
-                        })
-                    })
-                })
-            })
-            .unwrap_or_default()
-            .to_ascii_lowercase();
-    match status.as_str() {
-        "cancelled" | "canceled" | "interrupted" | "aborted" => OutcomeStatus::Cancelled,
-        "failed" | "error" => OutcomeStatus::Failed,
-        "replaced" => OutcomeStatus::Replaced,
-        _ => OutcomeStatus::Completed,
-    }
-}
-
-fn value_by_aliases<'a>(value: &'a Value, aliases: &[&str]) -> Option<&'a Value> {
-    aliases.iter().find_map(|alias| value.get(*alias))
-}
-
-fn value_string_by_aliases(value: Option<&Value>, aliases: &[&str]) -> Option<String> {
-    let value = value?;
-    value_by_aliases(value, aliases)
-        .and_then(|candidate| match candidate {
-            Value::String(text) => Some(text.clone()),
-            Value::Object(_) | Value::Array(_) => serde_json::to_string(candidate).ok(),
-            Value::Null => None,
-            other => Some(other.to_string()),
-        })
-        .filter(|text| !text.trim().is_empty())
-}
-
-fn is_assistant_or_reasoning_item(item: &Value) -> bool {
-    let item_type = value_string_by_aliases(Some(item), &["type", "kind"])
-        .unwrap_or_default()
-        .to_ascii_lowercase();
-    matches!(
-        item_type.as_str(),
-        "agentmessage" | "agent_message" | "assistantmessage" | "assistant_message" | "reasoning"
-    )
-}
-
-fn is_tool_item_type(item_type: &str) -> bool {
-    matches!(
-        item_type.to_ascii_lowercase().as_str(),
-        "commandexecution"
-            | "command_execution"
-            | "filechange"
-            | "file_change"
-            | "mcptoolcall"
-            | "mcp_tool_call"
-            | "toolcall"
-            | "tool_call"
-            | "dynamictoolcall"
-            | "dynamic_tool_call"
-            // Codex Responses often emits apply_patch as custom_tool_call (not fileChange).
-            | "customtoolcall"
-            | "custom_tool_call"
-            | "function_call"
-            | "functioncall"
-            | "apply_patch"
-            | "applypatch"
-    )
-}
-
-fn validate_owner(owner: &SharedRuntimeAttemptOwner) -> Result<(), String> {
-    for (field, value) in [
-        ("workspaceId", owner.workspace_id.as_str()),
-        ("providerRuntimeKey", owner.provider_runtime_key.as_str()),
-        ("sharedSessionId", owner.shared_session_id.as_str()),
-        ("sharedThreadId", owner.shared_thread_id.as_str()),
-        ("logicalTurnId", owner.logical_turn_id.as_str()),
-        ("attemptId", owner.attempt_id.as_str()),
-        ("bindingKey", owner.binding_key.as_str()),
-        ("bindingOperationId", owner.binding_operation_id.as_str()),
-    ] {
-        if value.trim().is_empty() {
-            return Err(format!("shared runtime owner {field} cannot be empty"));
-        }
-    }
-    if owner.shared_thread_id != format!("shared:{}", owner.shared_session_id) {
-        return Err("shared runtime owner session/thread identity mismatch".to_string());
-    }
-    if owner.execution_target_snapshot.engine != engine_token(owner.engine) {
-        return Err("shared runtime owner execution target engine mismatch".to_string());
-    }
-    if let Some(marker) = owner.context_marker.as_ref() {
-        if marker.package_id.trim().is_empty() || marker.source_checksum.trim().is_empty() {
-            return Err("shared runtime context marker cannot be empty".to_string());
-        }
-    }
-    Ok(())
-}
-
-fn same_durable_owner(left: &SharedRuntimeAttemptOwner, right: &SharedRuntimeAttemptOwner) -> bool {
-    left.workspace_id == right.workspace_id
-        && left.provider_runtime_key == right.provider_runtime_key
-        && left.shared_session_id == right.shared_session_id
-        && left.shared_thread_id == right.shared_thread_id
-        && left.logical_turn_id == right.logical_turn_id
-        && left.attempt_id == right.attempt_id
-        && left.binding_key == right.binding_key
-        && left.binding_operation_id == right.binding_operation_id
-        && left.engine == right.engine
-        && left.execution_target_snapshot == right.execution_target_snapshot
-        && left.context_marker == right.context_marker
-}
-
-fn insert_identity_owner(
-    index: &mut HashMap<RuntimeIdentityKey, String>,
-    key: RuntimeIdentityKey,
-    attempt_id: &str,
-) -> Result<(), String> {
-    if let Some(existing) = index.get(&key) {
-        if existing != attempt_id {
-            return Err(format!(
-                "shared runtime identity already owned by attempt {existing}"
-            ));
-        }
-    }
-    index.insert(key, attempt_id.to_string());
-    Ok(())
-}
-
-fn identity_key(owner: &SharedRuntimeAttemptOwner, identity: &str) -> RuntimeIdentityKey {
-    RuntimeIdentityKey {
-        workspace_id: owner.workspace_id.clone(),
-        engine: owner.engine,
-        provider_runtime_key: owner.provider_runtime_key.clone(),
-        identity: identity.to_string(),
-    }
-}
-
-fn runtime_scope_key(owner: &SharedRuntimeAttemptOwner) -> RuntimeScopeKey {
-    RuntimeScopeKey {
-        workspace_id: owner.workspace_id.clone(),
-        engine: owner.engine,
-        provider_runtime_key: owner.provider_runtime_key.clone(),
-    }
-}
-
-fn runtime_scope_key_for_ingress(ingress: &RuntimeIngress) -> RuntimeScopeKey {
-    RuntimeScopeKey {
-        workspace_id: ingress.workspace_id.clone(),
-        engine: ingress.engine,
-        provider_runtime_key: ingress.provider_runtime_key.clone(),
-    }
-}
-
-fn native_identity_key_for_ingress(ingress: &RuntimeIngress) -> Option<RuntimeIdentityKey> {
-    ingress
-        .native_session_id
-        .as_deref()
-        .map(|native_session_id| RuntimeIdentityKey {
-            workspace_id: ingress.workspace_id.clone(),
-            engine: ingress.engine,
-            provider_runtime_key: ingress.provider_runtime_key.clone(),
-            identity: native_session_id.to_string(),
-        })
-}
-
-fn normalize_identity(value: Option<&str>) -> Option<&str> {
-    value.map(str::trim).filter(|value| !value.is_empty())
-}
-
-fn normalize_native_session_identity(
-    engine: EngineType,
-    provider_runtime_key: Option<&str>,
-    value: Option<&str>,
-) -> Option<String> {
-    let normalized = normalize_identity(value)?;
-    // Claude / Kimi / Grok / OpenCode：catalog 与 FE hide 使用 `engine:{raw}`。
-    // Codex 保持 raw thread id（无前缀）。pending 占位原样保留，避免误写成
-    // `grok:grok-pending-shared-*`。
-    match engine {
-        EngineType::Claude
-        | EngineType::Kimi
-        | EngineType::Pi
-        | EngineType::Grok
-        | EngineType::OpenCode => {
-            let token = engine_token(engine);
-            let prefix = format!("{token}:");
-            if crate::shared_sessions::is_pending_shared_binding_thread_id(engine, normalized) {
-                return Some(normalized.to_string());
-            }
-            let raw = normalized
-                .strip_prefix(prefix.as_str())
-                .unwrap_or(normalized)
-                .trim();
-            if raw.is_empty() {
-                return None;
-            }
-            if crate::shared_sessions::is_pending_shared_binding_thread_id(engine, raw) {
-                return Some(raw.to_string());
-            }
-            Some(format!("{prefix}{raw}"))
-        }
-        EngineType::Qoder => {
-            if crate::shared_sessions::is_pending_shared_binding_thread_id(engine, normalized) {
-                return Some(normalized.to_string());
-            }
-            let provider_profile_id = provider_runtime_key.and_then(
-                crate::engine::qoder_provider_profile::qoder_provider_profile_id_from_runtime_key,
-            );
-            let identity =
-                crate::engine::qoder_provider_profile::parse_qoder_native_session_identity(
-                    normalized,
-                    provider_profile_id,
-                )
-                .ok()?;
-            // Runtime ingress 的 raw ACP session id 没有 distribution。只有明确
-            // 的 Qoder runtime key 才能把它升格为 durable Native identity；canonical
-            // identity 自带 profile，可用于兼容已经落盘的历史事件。
-            if provider_profile_id.is_none() && identity.is_legacy {
-                return None;
-            }
-            Some(identity.canonical_id())
-        }
-        EngineType::Codex | EngineType::Gemini | EngineType::Dsh => Some(normalized.to_string()),
-    }
-}
-
-pub(crate) fn is_missing_native_session_error(error: &str) -> bool {
-    let normalized = error.to_ascii_lowercase();
-    normalized.contains("no conversation found with session id")
-        || normalized.contains("conversation not found for session id")
-}
-
-fn engine_token(engine: EngineType) -> &'static str {
-    match engine {
-        EngineType::Claude => "claude",
-        EngineType::Codex => "codex",
-        EngineType::Gemini => "gemini",
-        EngineType::OpenCode => "opencode",
-        EngineType::Kimi => "kimi",
-        EngineType::Pi => "pi",
-        EngineType::Grok => "grok",
-        EngineType::Dsh => "dsh",
-        EngineType::Qoder => "qoder",
     }
 }
